fedora: Optimize Dockerfile
The docker.rpm file is actually docker-cd-17.03.0.ce-1.fc24.x86_64.rpm,
which is about the closest I found that still works with this version of
Fedora (hopefully the tests work with that too).  Same for the SELinux
package.  The old package failed due to an unresolved dependency on
docker-common.  Ideally, we'd use a package without any dependencies.